refactor: State 변경
- isMiss, isSecond 메소드 제거
- SecondaryState 추가
- 중복되는 메소드 제거
