+++ src/main/java/bowling/domain/state/First.java
@@ -28,16 +28,6 @@ public class First extends RunningState {
     }
 
     @Override
-    public Score score(Score score) {
-        return score.bowl(pin);
-    }
-
-    @Override
-    public boolean isDone() {
-        return false;
-    }
-
-    @Override
     public String toString() {
         return pin.toString();
     }
