+++ src/main/java/bowling/domain/state/Miss.java
@@ -4,18 +4,15 @@ import bowling.domain.score.Score;
 
 import static bowling.view.OutputView.COLUMN;
 
-public class Miss extends EndedState {
+public class Miss extends SecondaryState {
     private static final String SYMBOL = "-";
-
-    private final State previous;
 
     public Miss(State previous) {
         this(Pin.zero(), previous);
     }
 
     public Miss(Pin pin, State previous) {
-        super(pin);
-        this.previous = previous;
+        super(pin, previous);
     }
 
     @Override
@@ -26,27 +23,6 @@ public class Miss extends EndedState {
     @Override
     public Score score() {
         return previous.score().bowl(pin);
-    }
-
-    @Override
-    public Score score(Score score) {
-        Score nextScore = score.bowl(previous.pin);
-
-        if (nextScore.canScore()) {
-            return nextScore;
-        }
-
-        return nextScore.bowl(pin);
-    }
-
-    @Override
-    public boolean isDone() {
-        return true;
-    }
-
-    @Override
-    public boolean isMiss() {
-        return true;
     }
 
     @Override
