+++ src/main/java/bowling/domain/state/RunningState.java
@@ -4,4 +4,9 @@ public abstract class RunningState extends State {
     protected RunningState(Pin pin) {
         super(pin);
     }
+
+    @Override
+    public boolean isDone() {
+        return false;
+    }
 }
